Refactor to use array instead of list for instruction words
Refactored code to use a uint[] array instead of List<uint> for storing instruction words.
Initialized the `words` array with a size of `wordCount` and set the first element to `instructionStart`.
Adjusted the loop index to start from 1 and changed the method of adding elements to `words[i] = reader.ReadWord()`.