Remove `continue` statements in `Module` class loop
The `continue` statements have been removed from the `foreach` loop in the `Module` class within the `Module.cs` file. Previously, the loop would skip the remaining code within the loop body after adding an `instruction` to `debugInstructions` or `entryPoints`. This change allows further processing of the same `instruction` even if it has been added to one of these collections.